refactor(cmd): rename kong.cmd.utils.kill => kong.cmd.utils.process
There's a large refactor of this module in the pipeline, and renaming
from 'kill' to 'process' makes the API more sensible.
