+++ kong-3.5.0-0.rockspec
@@ -116,7 +116,7 @@ build = {
     ["kong.cmd.version"] = "kong/cmd/version.lua",
     ["kong.cmd.hybrid"] = "kong/cmd/hybrid.lua",
     ["kong.cmd.utils.log"] = "kong/cmd/utils/log.lua",
-    ["kong.cmd.utils.kill"] = "kong/cmd/utils/kill.lua",
+    ["kong.cmd.utils.process"] = "kong/cmd/utils/process.lua",
     ["kong.cmd.utils.env"] = "kong/cmd/utils/env.lua",
     ["kong.cmd.utils.migrations"] = "kong/cmd/utils/migrations.lua",
     ["kong.cmd.utils.tty"] = "kong/cmd/utils/tty.lua",
