+++ kong/cmd/quit.lua
@@ -1,7 +1,7 @@
 local nginx_signals = require "kong.cmd.utils.nginx_signals"
 local conf_loader = require "kong.conf_loader"
 local pl_path = require "pl.path"
-local kill = require "kong.cmd.utils.kill"
+local process = require "kong.cmd.utils.process"
 local log = require "kong.cmd.utils.log"
 
 local function execute(args)
@@ -24,7 +24,7 @@ local function execute(args)
     log.verbose("waiting %s seconds before quitting", args.wait)
     while twait > ngx.now() do
       ngx.sleep(0.2)
-      if not kill.is_running(conf.nginx_pid) then
+      if not process.is_running(conf.nginx_pid) then
         log.error("Kong stopped while waiting (unexpected)")
         return
       end
@@ -41,7 +41,7 @@ local function execute(args)
   local texp, running = tstart + math.max(args.timeout, 1) -- min 1s timeout
   repeat
     ngx.sleep(0.2)
-    running = kill.is_running(conf.nginx_pid)
+    running = process.is_running(conf.nginx_pid)
     ngx.update_time()
   until not running or ngx.now() >= texp
